Better schema validators class hierarchy + err messages
